polygonapi/requirements.txt
Django==2.2.3
djangorestframework==3.10.1
Shapely==1.8.2
